requirements.txt
Flask
gunicorn
requests
dash 
dash-renderer 
dash-core-components 
dash-html-components 
plotly 
pandas
dash_table
numpy
APScheduler
opencv-python-headless==4.2.0.32
pillow
matplotlib==3.3.3